fix: Update Information Highlight folder name for slash imports
