+++ modules/preview-react/index.ts
@@ -11,4 +11,4 @@ export * from './side-panel';
 export * from './status-indicator';
 export * from './text-area';
 export * from './text-input';
-export * from './InformationHighlight';
+export * from './information-highlight';
