+++ modules/preview-react/information-highlight/index.tsx
