+++ modules/preview-react/information-highlight/lib/InformationHighlight.tsx
