+++ modules/preview-react/information-highlight/lib/hooks/useInformationHighlightModel.ts
