+++ modules/preview-react/information-highlight/lib/parts/Body.tsx
