+++ modules/preview-react/information-highlight/lib/parts/Heading.tsx
