+++ modules/preview-react/information-highlight/spec/InformationHighlight.spec.tsx
