+++ modules/preview-react/information-highlight/stories/InformationHighlight.mdx
