+++ modules/preview-react/information-highlight/stories/InformationHighlight.stories.ts
@@ -2,7 +2,7 @@ import {Meta, StoryObj} from '@storybook/react';
 
 import mdxDoc from './InformationHighlight.mdx';
 
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 import {Basic as BasicExample} from './examples/Basic';
 import {Body as BodyExample} from './examples/Body';
 import {Heading as HeadingExample} from './examples/Heading';
