+++ modules/preview-react/information-highlight/stories/examples/Body.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 
 export const Body = () => {
   return (
