+++ modules/preview-react/information-highlight/stories/examples/Caution.tsx
@@ -1,30 +1,24 @@
 import React from 'react';
 import {Flex} from '@workday/canvas-kit-react/layout';
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 import {system} from '@workday/canvas-tokens-web';
 
-export const Critical = () => {
+export const Caution = () => {
   return (
     <Flex flexDirection={'column'} gap={system.space.x2}>
-      <InformationHighlight variant={'critical'}>
+      <InformationHighlight variant={'caution'}>
         <InformationHighlight.Icon />
-        <InformationHighlight.Heading>
-          {' '}
-          Attention! Highlight Something{' '}
-        </InformationHighlight.Heading>
+        <InformationHighlight.Heading> Caution: Highlight Something </InformationHighlight.Heading>
         <InformationHighlight.Body>
-          If you select the link below it will just reroute you back to this page
+          If you select the link below, nothing will happen
         </InformationHighlight.Body>
         <InformationHighlight.Link href="#hyperlink">View the Docs</InformationHighlight.Link>
       </InformationHighlight>
-      <InformationHighlight variant={'critical'} emphasis={'high'}>
+      <InformationHighlight variant={'caution'} emphasis={'high'}>
         <InformationHighlight.Icon />
-        <InformationHighlight.Heading>
-          {' '}
-          Attention! Highlight Something{' '}
-        </InformationHighlight.Heading>
+        <InformationHighlight.Heading> Caution: Highlight Something </InformationHighlight.Heading>
         <InformationHighlight.Body>
-          If you select the link below it will just reroute you back to this page
+          If you select the link below, nothing will happen
         </InformationHighlight.Body>
         <InformationHighlight.Link href="#hyperlink">View the Docs</InformationHighlight.Link>
       </InformationHighlight>
