+++ modules/preview-react/information-highlight/stories/examples/Critical.tsx
@@ -1,24 +1,30 @@
 import React from 'react';
 import {Flex} from '@workday/canvas-kit-react/layout';
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 import {system} from '@workday/canvas-tokens-web';
 
-export const Caution = () => {
+export const Critical = () => {
   return (
     <Flex flexDirection={'column'} gap={system.space.x2}>
-      <InformationHighlight variant={'caution'}>
+      <InformationHighlight variant={'critical'}>
         <InformationHighlight.Icon />
-        <InformationHighlight.Heading> Caution: Highlight Something </InformationHighlight.Heading>
+        <InformationHighlight.Heading>
+          {' '}
+          Attention! Highlight Something{' '}
+        </InformationHighlight.Heading>
         <InformationHighlight.Body>
-          If you select the link below, nothing will happen
+          If you select the link below it will just reroute you back to this page
         </InformationHighlight.Body>
         <InformationHighlight.Link href="#hyperlink">View the Docs</InformationHighlight.Link>
       </InformationHighlight>
-      <InformationHighlight variant={'caution'} emphasis={'high'}>
+      <InformationHighlight variant={'critical'} emphasis={'high'}>
         <InformationHighlight.Icon />
-        <InformationHighlight.Heading> Caution: Highlight Something </InformationHighlight.Heading>
+        <InformationHighlight.Heading>
+          {' '}
+          Attention! Highlight Something{' '}
+        </InformationHighlight.Heading>
         <InformationHighlight.Body>
-          If you select the link below, nothing will happen
+          If you select the link below it will just reroute you back to this page
         </InformationHighlight.Body>
         <InformationHighlight.Link href="#hyperlink">View the Docs</InformationHighlight.Link>
       </InformationHighlight>
