+++ modules/preview-react/information-highlight/stories/examples/Informational.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Flex} from '@workday/canvas-kit-react/layout';
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 import {system} from '@workday/canvas-tokens-web';
 
 export const Informational = () => {
