+++ modules/react/tokens/stories/Tokens.mdx
@@ -1,5 +1,5 @@
 import {ExampleCodeBlock} from '@workday/canvas-kit-docs';
-import {InformationHighlight} from '@workday/canvas-kit-preview-react/InformationHighlight';
+import {InformationHighlight} from '@workday/canvas-kit-preview-react/information-highlight';
 import * as TokenStories from './Tokens.stories';
 import {BorderRadius, Space, Depth, Colors, Type} from './examples/Tokens';
 import {Overview} from './examples/Overview';
